% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Shiny app for tourr D3 gui}
\usage{
launchApp(inputDataFile)
}
\arguments{
\item{inputDataFile}{Path to starting input data file (csv format)}
}
\description{
Shiny app for tourr D3 gui
}
\author{
Ursula Laa

Michael Kipp
}
